refactor: extract CI logs/await to github/ci_logs.py (#140)
