+++ tests/unit/github/test_github_await_workflow.py
@@ -45,7 +45,7 @@ class TestGitHubAwaitWorkflowCompletion:
         mock_client.get = AsyncMock(return_value=mock_run_response)
 
         with patch(
-            "src.mcp_server_git.github.workflows.github_client_context"
+            "src.mcp_server_git.github.ci_logs.github_client_context"
         ) as mock_context:
             mock_context.return_value.__aenter__.return_value = mock_client
 
@@ -116,7 +116,7 @@ class TestGitHubAwaitWorkflowCompletion:
         mock_client.get = AsyncMock(side_effect=mock_get)
 
         with patch(
-            "src.mcp_server_git.github.workflows.github_client_context"
+            "src.mcp_server_git.github.ci_logs.github_client_context"
         ) as mock_context:
             mock_context.return_value.__aenter__.return_value = mock_client
 
@@ -157,7 +157,7 @@ class TestGitHubAwaitWorkflowCompletion:
         mock_client.get = AsyncMock(return_value=mock_run_response)
 
         with patch(
-            "src.mcp_server_git.github.workflows.github_client_context"
+            "src.mcp_server_git.github.ci_logs.github_client_context"
         ) as mock_context:
             mock_context.return_value.__aenter__.return_value = mock_client
 
@@ -220,7 +220,7 @@ class TestGitHubAwaitWorkflowCompletion:
         mock_client.get = AsyncMock(side_effect=mock_get)
 
         with patch(
-            "src.mcp_server_git.github.workflows.github_client_context"
+            "src.mcp_server_git.github.ci_logs.github_client_context"
         ) as mock_context:
             mock_context.return_value.__aenter__.return_value = mock_client
 
@@ -280,7 +280,7 @@ class TestGitHubAwaitWorkflowCompletion:
         mock_client.get = AsyncMock(side_effect=mock_get)
 
         with patch(
-            "src.mcp_server_git.github.workflows.github_client_context"
+            "src.mcp_server_git.github.ci_logs.github_client_context"
         ) as mock_context:
             mock_context.return_value.__aenter__.return_value = mock_client
 
@@ -300,7 +300,7 @@ class TestGitHubAwaitWorkflowCompletion:
     async def test_authentication_error(self):
         """Test that authentication errors are handled properly."""
         with patch(
-            "src.mcp_server_git.github.workflows.github_client_context"
+            "src.mcp_server_git.github.ci_logs.github_client_context"
         ) as mock_context:
             mock_context.return_value.__aenter__.side_effect = ValueError(
                 "GitHub token not configured"
@@ -324,7 +324,7 @@ class TestGitHubAwaitWorkflowCompletion:
         mock_client.get = AsyncMock(return_value=mock_response)
 
         with patch(
-            "src.mcp_server_git.github.workflows.github_client_context"
+            "src.mcp_server_git.github.ci_logs.github_client_context"
         ) as mock_context:
             mock_context.return_value.__aenter__.return_value = mock_client
